refactor(phase-3h): extract JiraReportingService
Move filter and dashboard queries (get_filters, get_dashboards,
get_dashboard_details) out of JiraClient into a focused
JiraReportingService, continuing the ADR-002 god-class decomposition.

The service is exposed on JiraClient as self.reporting and the
client keeps thin delegators so existing call sites continue to
work unchanged. Like the other Phase 3 services this is HTTP-only
through the jira SDK session, so no Ruby-script escaping is involved.

The ``except (exceptions.HTTPError, AtlassianJIRAError)`` clause in
get_filters is rewritten as a multi-line tuple with a trailing
comma so ruff format preserves it.
